refactor: address daily quality findings for 2026-07-30
Share canvas add-widget catalog between dialog and empty-brush UI,
split empty-brush gesture detection from the popover, clarify shim vs
native OSC local title naming, centralize store OSC normalization, and
restore mention search result limit. Includes quality report score 80.
